requirements.txt
Flask==3.0.0
numpy==1.26.4
scikit-image==0.22.0
Pillow==10.2.0
matplotlib==3.8.2
PyPDF2==3.0.1
pandas==2.1.4
scikit-learn==1.3.2
